requirements.txt
openmeteo-requests
requests-cache 
retry-requests 
numpy 
pandas
jsonschema
scipy
scikit-learn
python-dotenv
fastapi
uvicorn
pydantic